Rename graphlink_licensing.py to graphlink_settings_store.py
The file has never had any licensing logic - it's two definitions,
_is_llama_cpp_gguf_path and SettingsManager, the persisted-settings
authority backend/settings.py wraps. The name actively misleads anyone
searching the repo for real licensing/entitlement logic (there is none)
or for where settings persistence lives (it's here).

Filename-only rename: SettingsManager's own name, its public API, and
every behavior are unchanged. Updated all 13 import/reference sites
across 10 files plus 5 sites in GRAPHLINK_REPO_NAVIGATION.md, and moved
pyproject.toml's py-modules entry to keep that list's alphabetical
ordering intact rather than swapping the string in place.

Verified beyond the test suite: built the real wheel, installed it into
a clean venv, and confirmed graphlink_settings_store imports correctly
while graphlink_licensing is gone, from outside the repo checkout - the
same class of check that would have caught the R7 wheel bug fixed
earlier in this stabilization pass.
