+++ pyproject.toml
@@ -85,7 +85,6 @@ py-modules = [
     "graphlink_chat_agent",
     "graphlink_desktop",
     "graphlink_grid_view_settings",
-    "graphlink_licensing",
     "graphlink_memory",
     "graphlink_model_catalog",
     "graphlink_navigation_pins",
@@ -93,6 +92,7 @@ py-modules = [
     "graphlink_process_env",
     "graphlink_prompts",
     "graphlink_secrets",
+    "graphlink_settings_store",
     "graphlink_task_config",
     "graphlink_token_estimator",
     "graphlink_version",
